move getGameById into own file
